Create AssetSelector component
