+++ src/components/AssetSelector/index.tsx
@@ -1,0 +1,46 @@
+import { Asset } from "stellar-sdk";
+import { stringifyStellarAsset } from "../../helpers/stellar";
+import LabelledSelector from "../LabelledSelector";
+import { h } from "preact";
+
+interface AssetSelectorProps {
+  selectedAsset?: Asset;
+  onChange: (asset: Asset) => void;
+  assets: Asset[];
+  style?: React.CSSProperties;
+}
+
+function AssetSelector(props: AssetSelectorProps): JSX.Element {
+  const { assets, selectedAsset } = props;
+
+  const items = assets.map((asset) => {
+    return {
+      displayName: asset.getCode(),
+      id: stringifyStellarAsset(asset),
+    };
+  });
+
+  const selectedAssetItem = selectedAsset
+    ? {
+        displayName: selectedAsset.getCode(),
+        id: stringifyStellarAsset(selectedAsset),
+      }
+    : undefined;
+
+  return (
+    <LabelledSelector
+      items={items}
+      label="Asset"
+      onChange={(newItem) => {
+        const newAsset = assets.find((asset) => {
+          return stringifyStellarAsset(asset) === newItem.id;
+        });
+        newAsset && props.onChange(newAsset);
+      }}
+      value={selectedAssetItem}
+      style={props.style}
+    />
+  );
+}
+
+export default AssetSelector;
